CBORSerializer impl
