+++ _benchmarks/suites/table_query_suite_with_table_serializer.go
@@ -6,6 +6,7 @@ import (
 
 	"github.com/go-bond/bond"
 	"github.com/go-bond/bond/_benchmarks/bench"
+	"github.com/go-bond/bond/serializers"
 	"github.com/vmihailenco/msgpack/v5"
 )
 
@@ -24,8 +25,10 @@ func BenchmarkTableQueryWithTableSerializerSuite(bs *bench.BenchmarkSuite) []ben
 		Name       string
 		Serializer bond.Serializer[any]
 	}{
-		//{"JSONSerializer", &bond.JsonSerializer{}},
-		{"MsgpackSerializer", &bond.MsgpackSerializer{}},
+		{"JSONSerializer", &serializers.JsonSerializer{}},
+		{"MsgpackSerializer", &serializers.MsgpackSerializer{}},
+		{"MsgpackGenSerializer", &serializers.MsgpackGenSerializer{}},
+		{"CBORSerializer", &serializers.CBORSerializer{}},
 	}
 
 	var results []bench.BenchmarkResult
