+++ _benchmarks/suites/table_upsert_suite.go
@@ -8,6 +8,7 @@ import (
 	"github.com/cockroachdb/pebble"
 	"github.com/go-bond/bond"
 	"github.com/go-bond/bond/_benchmarks/bench"
+	"github.com/go-bond/bond/serializers"
 	"github.com/vmihailenco/msgpack/v5"
 )
 
@@ -26,8 +27,10 @@ func BenchmarkTableUpsertSuite(bs *bench.BenchmarkSuite) []bench.BenchmarkResult
 		Name       string
 		Serializer bond.Serializer[any]
 	}{
-		//{"JSONSerializer", &bond.JsonSerializer{}},
-		{"MsgpackSerializer", &bond.MsgpackSerializer{}},
+		{"JSONSerializer", &serializers.JsonSerializer{}},
+		{"MsgpackSerializer", &serializers.MsgpackSerializer{}},
+		{"MsgpackGenSerializer", &serializers.MsgpackGenSerializer{}},
+		{"CBORSerializer", &serializers.CBORSerializer{}},
 	}
 
 	var results []bench.BenchmarkResult
