+++ helpers.go
@@ -55,28 +55,3 @@ type UUIDGenerator struct {
 func (n *UUIDGenerator) Next() (uuid.UUID, error) {
 	return uuid.New(), nil
 }
-
-type Lazy[T any] struct {
-	getFunc func() (T, error)
-}
-
-func (l Lazy[T]) Get() (T, error) {
-	return l.getFunc()
-}
-
-type SyncPool[T any] interface {
-	Get() T
-	Put(T)
-}
-
-type SyncPoolWrapper[T any] struct {
-	sync.Pool
-}
-
-func (s *SyncPoolWrapper[T]) Get() T {
-	return s.Pool.Get().(T)
-}
-
-func (s *SyncPoolWrapper[T]) Put(t T) {
-	s.Pool.Put(t)
-}
