+++ lazy.go
@@ -1,0 +1,9 @@
+package bond
+
+type Lazy[T any] struct {
+	GetFunc func() (T, error)
+}
+
+func (l Lazy[T]) Get() (T, error) {
+	return l.GetFunc()
+}
